+++ features/mpcredits/src/main/java/com/jpp/mpcredits/CreditsViewModel.kt
@@ -50,8 +50,8 @@ class CreditsViewModel @Inject constructor(
     init {
         _viewState.addSource(creditsInteractor.events) { event ->
             when (event) {
-                is NotConnectedToNetwork -> _viewState.value = CreditsViewState.showNoConnectivityError(retry)
-                is UnknownError -> _viewState.value = CreditsViewState.showUnknownError(retry)
+                is NotConnectedToNetwork -> _viewState.value = _viewState.value?.showNoConnectivityError(retry)
+                is UnknownError -> _viewState.value = _viewState.value?.showUnknownError(retry)
                 is Success -> mapCreditsAndPushViewState(event.credits)
                 is AppLanguageChanged -> refreshCreditsData(currentParam.movieId)
             }
@@ -126,11 +126,11 @@ class CreditsViewModel @Inject constructor(
      */
     private fun mapCreditsAndPushViewState(credits: Credits) {
         when (credits.cast.isEmpty() && credits.crew.isEmpty()) {
-            true -> _viewState.value = CreditsViewState.showNoCreditsAvailable()
+            true -> _viewState.value = _viewState.value?.showNoCreditsAvailable()
             false -> {
                 viewModelScope.launch {
                     withContext(ioDispatcher) {
-                        CreditsViewState.showCredits(
+                        _viewState.value?.showCredits(
                                 credits.cast
                                         .map { imagesPathInteractor.configureCastCharacter(currentParam.targetImageSize, it) }
                                         .map { mapCastCharacterToCreditPerson(it) }
